requirements.txt
discord.py
requests
threading